Combine DataStore.cs into ObservationsAndNotesViewModel.cs. Remove DataStore.cs.
