+++ ERIS.Mobile/ERIS.Mobile/ViewModels/ObservationsAndNotesViewModel.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ERIS.Mobile.Services;
+using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Windows.Input;
@@ -8,16 +9,27 @@ namespace ERIS.Mobile.ViewModels
 {
     public class ObservationsAndNotesViewModel : AssessmentDetailsUpdater
     {
+        private SendData send;
+        public ICommand checkSubmitButton { get; }
         public ICommand observationsAndNotesUnfocused { get; }
+
         public ObservationsAndNotesViewModel()
         {
             observationsAndNotesUnfocused = new Command<FocusEventArgs>(SetObservationsAndNotes);
+            
+            send = new SendData();
+            
+            checkSubmitButton = new Command(SubmitPressed);
         }
         
-
         private void SetObservationsAndNotes(FocusEventArgs args)
         {
             SetAssessmentDetailsStringAndUpdateJsonFile(nameof(assessmentDetails.ObservationsAndNotes), ((Editor)(args.VisualElement)));
+        }
+
+        private void SubmitPressed()
+        {
+            send.PostDetails();
         }
 
         public string ObservationsAndNotes
