+++ ERIS.Mobile/ERIS.Mobile/Views/ObservationsAndNotesPage.xaml
@@ -2,7 +2,6 @@
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" xmlns:xct="http://xamarin.com/schemas/2020/toolkit"
              x:Class="ERIS.Mobile.Views.ObservationsAndNotesPage"
-             xmlns:vm="clr-namespace:ERIS.Mobile.ViewModels" x:DataType="vm:DataStore"
              BackgroundColor="{StaticResource AssessmentBackgroundColor}">
     <ContentPage.Content>
         <StackLayout Style="{StaticResource MainStackLayout}">
