+++ ERIS.Mobile/ERIS.Mobile/Views/ObservationsAndNotesPage.xaml.cs
@@ -4,10 +4,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
-using ERIS.Mobile.ViewModels;
 
 namespace ERIS.Mobile.Views
 {
